fix(frontend): move gamme-autolink to app/utils (owned glob) + clear detector FP
CI fixes on top of the iOS lookbehind crash fix:

- Move app/lib/gamme-autolink.ts -> app/utils/gamme-autolink.ts. app/utils/** is
  already owner+domain-mapped (D8 / @ak125/frontend-team) so the "Block-new" gate
  passes without a new ownership.yaml entry; it also co-locates with the other
  HTML-string utils (editorial-parser, sanitize-advice). Imports updated.

- Reword a tokenizer comment so the `impeccable` design detector no longer flags
  an example `<img …>` written inside a code comment as a broken-image (false
  positive). Detector total back to baseline (9).

41 tests green; tsc + eslint + prettier clean.
