+++ frontend/app/components/pieces/ConseilsSection.tsx
@@ -9,8 +9,8 @@ import {
 } from "lucide-react";
 import { useState, useMemo, memo } from "react";
 
-import { addGammeLinks } from "~/lib/gamme-autolink";
 import { pluralizePieceName } from "~/lib/seo-utils";
+import { addGammeLinks } from "~/utils/gamme-autolink";
 import { HtmlContent } from "../seo/HtmlContent";
 
 interface ConseilItem {
@@ -115,7 +115,7 @@ function categorizeConseil(title: string, content: string): string {
 }
 
 /** Tailwind classes for gamme auto-links inside conseil HTML.
- *  Auto-linking logic lives in the shared, iOS-safe `~/lib/gamme-autolink` helper. */
+ *  Auto-linking logic lives in the shared, iOS-safe `~/utils/gamme-autolink` helper. */
 const CONSEIL_GAMME_LINK_CLASS =
   "text-green-600 hover:text-green-800 underline decoration-dotted hover:decoration-solid font-medium";
